requirements/base.txt
Flask
gunicorn
pytest
python-dotenv